requirements.txt
discord.py==2.4.0
audioop-lts>=0.2.1  # Python 3.13+ 호환성
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
sqlalchemy==2.0.36
apscheduler==3.10.4
matplotlib>=3.10.0
python-dotenv==1.0.1
